chore: update docs of model package classes
