+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/api/exception/api_error_response_exception.dart
@@ -1,58 +1,22 @@
 part of 'api_exception.dart';
 
+/// Api did response but status code is not >=200 <300.
 final class ApiErrorResponseException extends ApiException {
-  /// The client has not sent valid request, could be an invalid HTTP in
-  /// general or provided not correct headers, path or query arguments.
-  static const int badRequest = 400;
-
-  /// The client has not sent valid authentication credentials for the
-  /// requested resource.
-  static const int unauthorized = 401;
-
-  /// The client has not sent valid authentication credentials for the
-  /// requested resource.
-  static const int forbidden = 403;
-
-  /// In context of request.
-  static const int notFound = 404;
-
-  /// A request could not be completed due to a conflict
-  /// with the current state of the target resource.
-  static const int conflict = 409;
-
-  /// The client has not sent valid data in its request, headers, parameters
-  /// or body.
-  static const int preconditionFailed = 412;
-
-  /// The client sent a request with the URI is longer than the server is
-  /// willing to interpret
-  static const int uriTooLong = 414;
-
-  /// The client has sent too many requests in a given amount of time.
-  static const int tooManyRequests = 429;
-
-  /// The client sent a request with too large header fields.
-  static const int requestHeaderFieldsTooLarge = 431;
-
-  /// An internal server error occurred.
-  ///
-  /// The contents of this response should be reported to the projects
-  /// issue tracker.
-  static const int internalServerError = 500;
-
-  /// The service is not available, try again later.
-  ///
-  /// This is returned when the service either has not started, or has become
-  /// unavailable.
-  static const int serviceUnavailable = 503;
-
+  /// Http response code
   final int statusCode;
+
+  /// Optional response body.
+  ///
+  /// It's dynamic changes base on endpoint and status code. That's why it's a [Object].
   final Object? error;
 
   const ApiErrorResponseException({
     required this.statusCode,
     this.error,
   });
+
+  const ApiErrorResponseException.notFound([this.error])
+      : statusCode = ApiResponseStatusCode.notFound;
 
   @override
   List<Object?> get props => [statusCode, error];
