+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/api/exception/api_malformed_body_exception.dart
@@ -1,5 +1,6 @@
 part of 'api_exception.dart';
 
+/// Response body did not match what was expected.
 final class ApiMalformedBodyException extends ApiException {
   const ApiMalformedBodyException();
