+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/auth/rbac_token.dart
@@ -1,2 +1,4 @@
 /// An entity which represents authorization token.
+///
+/// Its simple wrapper to enforce strong type instead of just using [String].
 extension type const RbacToken(String value) {}
