+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/catalyst_voices_models.dart
@@ -1,12 +1,12 @@
 library catalyst_voices_models;
 
+export 'api/api_response_status_code.dart';
 export 'api/exception/api_exception.dart';
 export 'auth/password_strength.dart';
 export 'auth/rbac_token.dart';
 export 'auth/seed_phrase.dart';
 export 'blockchain/blockchain_slot_config.dart';
 export 'campaign/campaign.dart';
-export 'campaign/campaign_base.dart';
 export 'campaign/campaign_category.dart';
 export 'campaign/campaign_timeline.dart';
 export 'campaign/current_campaign.dart';
