+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/config/env_vars/dart_define_env_vars.dart
@@ -1,9 +1,11 @@
+import 'package:catalyst_voices_models/catalyst_voices_models.dart';
 import 'package:catalyst_voices_models/src/config/env_vars/env_vars.dart';
 import 'package:flutter/services.dart';
 
 /// Have to mach against --dart-define variables
 const _envNameKey = 'ENV_NAME';
 
+/// Returns [AppEnvironmentType] settings from Dart's environment variables.
 EnvVars getDartEnvVars() {
   const envName = bool.hasEnvironment(_envNameKey) ? String.fromEnvironment(_envNameKey) : null;
