+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/errors/errors.dart
@@ -1,8 +1,7 @@
 export 'crypto_exception.dart';
 export 'email_already_used_exception.dart';
-export 'network_error.dart';
 export 'not_found_exception.dart';
 export 'resource_conflict_exception.dart';
-export 'secure_storage_error.dart';
+export 'sync_exception.dart';
 export 'unauthorized_exception.dart';
 export 'vault_exception.dart';
