--- catalyst_voices/packages/internal/catalyst_voices_models/lib/src/errors/secure_storage_error.dart
@@ -1,16 +1,0 @@
-enum SecureStorageError implements Comparable<SecureStorageError> {
-  canNotReadData('Cannot Read Data From Secure Storage'),
-  canNotSaveData('Cannot Save Data From Secure Storage');
-
-  final String description;
-
-  const SecureStorageError(this.description);
-
-  @override
-  int compareTo(SecureStorageError other) {
-    return description.compareTo(other.description);
-  }
-
-  @override
-  String toString() => 'SecureStorageError(description: $description)';
-}
