+++ catalyst_voices/packages/internal/catalyst_voices_models/lib/src/hi_lo/hi_lo.dart
@@ -1,5 +1,6 @@
 import 'package:equatable/equatable.dart';
 
+/// Abstract HiLo pattern representation.
 abstract base class HiLo<T extends Object> extends Equatable {
   final T high;
   final T low;
