+++ catalyst_voices/packages/internal/catalyst_voices_view_models/test/campaign/campaign_stage_test.dart
@@ -14,6 +14,7 @@ void main() {
       endDate: date,
       proposalsCount: 0,
       publish: CampaignPublish.draft,
+      categoriesCount: 0,
     );
 
     test('draft campaign resolves to draft stage', () {
